rename guui components called index
