+++ packages/guui/components/CloseButton/CloseButton.tsx
