+++ packages/guui/index.ts
@@ -1,6 +1,6 @@
 // Components
-export { default as CloseButton } from './components/CloseButton';
-export { default as Dropdown } from './components/Dropdown';
-export { default as BigNumber } from './components/BigNumber';
-export { default as Container } from './components/Container';
-export { Cols as GridCols, Row as GridRow } from './components/Grid';
+export { default as CloseButton } from './components/CloseButton/CloseButton';
+export { default as Dropdown } from './components/Dropdown/Dropdown';
+export { default as BigNumber } from './components/BigNumber/BigNumber';
+export { default as Container } from './components/Container/Container';
+export { Cols as GridCols, Row as GridRow } from './components/Grid/Grid';
